add fetch candidates to crud
