+++ geoparser/annotator/annotator.py
@@ -17,135 +17,6 @@ class GeoparserAnnotator(Geoparser):
         self.nlp = None
         self.transformer = None
 
-    def get_existing_loc_id(self, toponym: dict) -> str:
-        return toponym.get("loc_id", "")
-
-    def get_candidate_descriptions(
-        self, toponym: dict, toponym_text: str, query_text: str
-    ) -> tuple[list[dict], bool]:
-
-        # Get coordinate columns and CRS from gazetteer config
-        coord_config = self.gazetteer.config.location_coordinates
-        x_col = coord_config.x_column
-        y_col = coord_config.y_column
-        crs = coord_config.crs
-
-        # Prepare coordinate transformer if needed
-        if crs != "EPSG:4326":
-            # Define transformer to WGS84
-            coord_transformer = Transformer.from_crs(crs, "EPSG:4326", always_xy=True)
-        else:
-            coord_transformer = None
-
-        # Use query_text if provided, else use toponym_text
-        search_text = query_text if query_text else toponym_text
-
-        # Get candidate IDs and locations based on the search_text
-        candidates = self.gazetteer.query_candidates(search_text)
-        candidate_locations = self.gazetteer.query_locations(candidates)
-
-        # Prepare candidate descriptions and attributes
-        candidate_descriptions = []
-        for location in candidate_locations:
-            description = self.gazetteer.get_location_description(location)
-
-            # Get coordinates
-            x = location.get(x_col)
-            y = location.get(y_col)
-
-            if x is not None and y is not None:
-                try:
-                    x = float(x)
-                    y = float(y)
-                    if coord_transformer:
-                        lon, lat = coord_transformer.transform(x, y)
-                    else:
-                        lon, lat = x, y
-                except (ValueError, TypeError):
-                    lat = None
-                    lon = None
-            else:
-                lat = None
-                lon = None
-
-            candidate_descriptions.append(
-                {
-                    "loc_id": location[self.gazetteer.config.location_identifier],
-                    "description": description,
-                    "attributes": location,  # Include all attributes for filtering
-                    "latitude": lat,
-                    "longitude": lon,
-                }
-            )
-
-        existing_loc_id = self.get_existing_loc_id(toponym)
-
-        append_existing_candidate = (
-            bool(existing_loc_id) and not existing_loc_id in candidates
-        )
-
-        if append_existing_candidate:
-            existing_location = self.gazetteer.query_locations([existing_loc_id])[0]
-            existing_description = self.gazetteer.get_location_description(
-                existing_location
-            )
-
-            # Get coordinates
-            x = existing_location.get(x_col)
-            y = existing_location.get(y_col)
-
-            if x is not None and y is not None:
-                try:
-                    x = float(x)
-                    y = float(y)
-                    if coord_transformer:
-                        lon, lat = coord_transformer.transform(x, y)
-                    else:
-                        lon, lat = x, y
-                except (ValueError, TypeError):
-                    lat = None
-                    lon = None
-            else:
-                lat = None
-                lon = None
-
-            existing_annotation = {
-                "loc_id": existing_loc_id,
-                "description": existing_description,
-                "attributes": existing_location,
-                "latitude": lat,
-                "longitude": lon,
-            }
-
-            candidate_descriptions.append(existing_annotation)
-        return candidate_descriptions, append_existing_candidate
-
-    def get_filter_attributes(self) -> list[str]:
-        location_identifier = self.gazetteer.config.location_identifier
-        location_columns = self.gazetteer.config.location_columns
-
-        filter_attributes = [
-            col.name
-            for col in location_columns
-            if col.type == "TEXT"
-            and col.name != location_identifier
-            and not col.name.endswith(location_identifier)
-        ]
-        return filter_attributes
-
-    def get_candidates(self, toponym: dict, toponym_text: str, query_text: str) -> dict:
-        candidate_descriptions, existing_candidate_is_appended = (
-            self.get_candidate_descriptions(toponym, toponym_text, query_text)
-        )
-        return {
-            "candidates": candidate_descriptions,
-            "filter_attributes": self.get_filter_attributes(),
-            "existing_loc_id": self.get_existing_loc_id(toponym),
-            "existing_candidate": (
-                candidate_descriptions[-1] if existing_candidate_is_appended else None
-            ),
-        }
-
     def annotate_toponyms(
         self,
         toponyms: list[dict],
